新增 - 【考勤打卡】查询归档报表表头 - 【考勤打卡】查询所有归档规则 - 【飞书人事（企业版）】通过/拒绝审批任务 - 【应用信息】获取消息推送概览
变更
- 查询统计数据 响应体
- 新建招聘官网投递 请求体

+++ src/Hire/GetHireV1TalentsResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-26
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-26
+// Last Modified On : 2024-07-24
 // ************************************************************************
 // <copyright file="GetHireV1TalentsResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -281,7 +281,7 @@ public record GetHireV1TalentsResponseDto
             /// <para>示例值：293016767159</para>
             /// </summary>
             [JsonPropertyName("birthday")]
-            public int? Birthday { get; set; }
+            public long? Birthday { get; set; }
 
             /// <summary>
             /// <para>创建人</para>
